Fix -i flag behavior
If isoform filtering was not desired (via the -i flag), the transcript data structure was left unmodified, which did not match the output format of the longest-isoform processing function. Now, if filtering is not desired the data is still processed to conform to the output structure of the filtering process.