Allow reciprologs to be found between >2 files
Now accepts more than two files as input, and will report reciprologs in groups of varying sizes.

Operates on the assumption that if A is a reciprolog of B, and B is a reciprolog of C, then A is also a reciprolog of C. This may be further developed in the future to enable restricting to cases where ABC are all pairwise reciprologs of one another (although what difference this would make biologically is unclear).